move new core code to the only module where it's used
core must be as lean as possible
