+++ modules/study/src/main/model.scala
@@ -21,3 +21,13 @@ case class AddNode(
     opts: MoveOpts,
     relay: Option[Chapter.Relay] = None
 )(using val who: Who)
+
+enum StudyGroup:
+  case all, mine, mineMember, minePublic, minePrivate, mineLikes, byOwner, staffPicks, search
+  case topic(topicName: Option[String])
+  def isTopic: Boolean = this match
+    case StudyGroup.topic(_) => true
+    case _ => false
+  def isPersonal: Boolean = this match
+    case StudyGroup.mine | StudyGroup.mineMember | StudyGroup.minePublic | StudyGroup.minePrivate => true
+    case _ => false
